Python-FastApi/pyproject.toml
[project]
name = "python-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
